refactor: Extract SwiftUI components to fix SwiftLint violations
- Extract BeamView and BeamGroup components to BeamView.swift
- Extract DrumBeatView and related components to DrumBeatView.swift
- Extract music notation components to MusicNotationViews.swift
- Extract beam grouping logic to BeamGroupingLogic.swift
- Reduce GameplayView from 701 lines to 365 lines (under 400 line limit)
- Reduce type body length from 315+ lines to 274 lines
- Eliminate serious SwiftLint violations while maintaining functionality
- Improve code organization and modularity
