+++ Virgo/views/GameplayView.swift
@@ -197,7 +197,7 @@ struct GameplayView: View {
     
     // MARK: - Drum Notation
     private func drumNotationView(measurePositions: [GameplayLayout.MeasurePosition]) -> some View {
-        let beamGroups = calculateBeamGroups(from: drumBeats)
+        let beamGroups = BeamGroupingHelper.calculateBeamGroups(from: drumBeats)
         
         return ZStack {
             // Render beams first (behind notes)
@@ -290,59 +290,6 @@ struct GameplayView: View {
         .background(Color.black.opacity(0.9))
     }
     
-    // MARK: - Beam Grouping Logic
-    private func calculateBeamGroups(from beats: [DrumBeat]) -> [BeamGroup] {
-        var beamGroups: [BeamGroup] = []
-        var currentGroup: [DrumBeat] = []
-        
-        for beat in beats {
-            // Only group eighth notes and shorter that need flags
-            if beat.interval.needsFlag {
-                // Check if this beat is consecutive to the previous one
-                if let lastBeat = currentGroup.last {
-                    let timeDifference = abs(beat.timePosition - lastBeat.timePosition)
-                    let measureNumber = Int(beat.timePosition)
-                    let lastMeasureNumber = Int(lastBeat.timePosition)
-                    
-                    // Group notes if they're in the same measure and consecutive (within 0.3 time units)
-                    if measureNumber == lastMeasureNumber && timeDifference <= 0.3 {
-                        currentGroup.append(beat)
-                    } else {
-                        // Finish current group if it has 2+ notes
-                        if currentGroup.count >= 2 {
-                            beamGroups.append(BeamGroup(
-                                id: "beam_\(currentGroup.first?.id ?? 0)",
-                                beats: currentGroup
-                            ))
-                        }
-                        currentGroup = [beat]
-                    }
-                } else {
-                    currentGroup = [beat]
-                }
-            } else {
-                // Finish current group for non-beamable notes
-                if currentGroup.count >= 2 {
-                    beamGroups.append(BeamGroup(
-                        id: "beam_\(currentGroup.first?.id ?? 0)",
-                        beats: currentGroup
-                    ))
-                }
-                currentGroup = []
-            }
-        }
-        
-        // Don't forget the last group
-        if currentGroup.count >= 2 {
-            beamGroups.append(BeamGroup(
-                id: "beam_\(currentGroup.first?.id ?? 0)",
-                beats: currentGroup
-            ))
-        }
-        
-        return beamGroups
-    }
-    
     // MARK: - Computed Properties
     private var drumBeats: [DrumBeat] {
         // If track has no notes, create some default beats for demonstration
@@ -410,289 +357,6 @@ struct GameplayView: View {
         if isPlaying {
             startPlayback()
         }
-    }
-}
-
-// MARK: - Supporting Models
-struct DrumBeat {
-    let id: Int
-    let drums: [DrumType]
-    let timePosition: Double
-    let interval: NoteInterval
-}
-
-// MARK: - Beam Group Models
-struct BeamGroup: Identifiable {
-    let id: String
-    let beats: [DrumBeat]
-    
-    var beamCount: Int {
-        return beats.map { $0.interval.flagCount }.max() ?? 1
-    }
-}
-
-// MARK: - Beam Group View
-struct BeamGroupView: View {
-    let beamGroup: BeamGroup
-    let measurePositions: [GameplayLayout.MeasurePosition]
-    let timeSignature: TimeSignature
-    let isActive: Bool
-    
-    var body: some View {
-        ZStack {
-            ForEach(0..<beamGroup.beamCount, id: \.self) { beamLevel in
-                BeamView(
-                    beats: beamGroup.beats,
-                    beamLevel: beamLevel,
-                    measurePositions: measurePositions,
-                    timeSignature: timeSignature,
-                    isActive: isActive
-                )
-            }
-        }
-    }
-}
-
-// MARK: - Individual Beam View
-struct BeamView: View {
-    let beats: [DrumBeat]
-    let beamLevel: Int
-    let measurePositions: [GameplayLayout.MeasurePosition]
-    let timeSignature: TimeSignature
-    let isActive: Bool
-    
-    var body: some View {
-        // Filter beats that need this beam level
-        let beamedBeats = beats.filter { $0.interval.flagCount > beamLevel }
-        
-        guard beamedBeats.count >= 2,
-              let firstBeat = beamedBeats.first,
-              let lastBeat = beamedBeats.last else {
-            return AnyView(EmptyView())
-        }
-        
-        // Calculate beam positions
-        let firstMeasureIndex = firstBeat.id / 1000
-        let lastMeasureIndex = lastBeat.id / 1000
-        
-        guard let firstMeasurePos = measurePositions.first(where: { $0.measureIndex == firstMeasureIndex }),
-              let lastMeasurePos = measurePositions.first(where: { $0.measureIndex == lastMeasureIndex }),
-              firstMeasurePos.row == lastMeasurePos.row else {
-            return AnyView(EmptyView())
-        }
-        
-        // Calculate X positions
-        let firstBeatOffset = firstBeat.timePosition - Double(firstMeasureIndex)
-        let lastBeatOffset = lastBeat.timePosition - Double(lastMeasureIndex)
-        let firstBeatIndex = Int(firstBeatOffset * Double(timeSignature.beatsPerMeasure))
-        let lastBeatIndex = Int(lastBeatOffset * Double(timeSignature.beatsPerMeasure))
-        
-        let startX = GameplayLayout.noteXPosition(measurePosition: firstMeasurePos, beatIndex: firstBeatIndex, timeSignature: timeSignature) + 7 // Stem offset
-        let endX = GameplayLayout.noteXPosition(measurePosition: lastMeasurePos, beatIndex: lastBeatIndex, timeSignature: timeSignature) + 7
-        
-        // Beam Y position (above the notes, accounting for beam level)
-        let baseY = GameplayLayout.StaffLinePosition.line3.absoluteY(for: firstMeasurePos.row) - 60 - CGFloat(beamLevel * 6)
-        
-        return AnyView(
-            Path { path in
-                path.move(to: CGPoint(x: startX, y: baseY))
-                path.addLine(to: CGPoint(x: endX, y: baseY))
-            }
-            .stroke(isActive ? Color.yellow : Color.white, lineWidth: 4)
-        )
-    }
-}
-
-// MARK: - Drum Beat View
-struct DrumBeatView: View {
-    let beat: DrumBeat
-    let isActive: Bool
-    let row: Int
-    let isBeamed: Bool
-    
-    // Calculate stem connection info for multiple simultaneous notes
-    private var stemInfo: StemConnectionInfo {
-        guard beat.interval.needsStem && beat.drums.count > 1 else {
-            return StemConnectionInfo(hasConnectedStem: false, stemTop: 0, stemBottom: 0, mainStemX: 7)
-        }
-        
-        let noteOffsets = beat.drums.map { $0.notePosition.yOffset }
-        let topOffset = noteOffsets.min() ?? 0 // Highest note (smallest y offset)
-        let bottomOffset = noteOffsets.max() ?? 0 // Lowest note (largest y offset)
-        
-        // Only connect stems if notes span multiple staff positions
-        let shouldConnect = abs(topOffset - bottomOffset) > GameplayLayout.staffLineSpacing
-        
-        if shouldConnect {
-            // Extend stem upward from the topmost note only (standard music notation)
-            let stemExtension: CGFloat = 35
-            return StemConnectionInfo(
-                hasConnectedStem: true,
-                stemTop: topOffset - stemExtension,
-                stemBottom: bottomOffset, // Don't extend beyond the bottommost note
-                mainStemX: 7
-            )
-        } else {
-            return StemConnectionInfo(hasConnectedStem: false, stemTop: 0, stemBottom: 0, mainStemX: 7)
-        }
-    }
-    
-    var body: some View {
-        ZStack {
-            // Beat column background
-            Rectangle()
-                .frame(width: 30, height: GameplayLayout.staffHeight)
-                .foregroundColor(isActive ? Color.purple.opacity(0.3) : Color.clear)
-                .cornerRadius(4)
-            
-            // Connected stem for multiple notes (if applicable)
-            if stemInfo.hasConnectedStem && beat.interval.needsStem {
-                let stemTop: CGFloat = isBeamed ? -60 : stemInfo.stemTop  // End at beam if beamed
-                let stemBottom = stemInfo.stemBottom
-                let stemHeight = stemBottom - stemTop
-                let stemCenterY = (stemTop + stemBottom) / 2
-                
-                Rectangle()
-                    .frame(width: 2, height: stemHeight)
-                    .foregroundColor(isActive ? .yellow : .white)
-                    .offset(x: stemInfo.mainStemX, y: stemCenterY)
-            }
-            
-            // Flags on the main stem (for connected stems, only if not beamed)
-            if stemInfo.hasConnectedStem && beat.interval.needsFlag && !isBeamed {
-                ForEach(0..<beat.interval.flagCount, id: \.self) { flagIndex in
-                    FlagView(flagIndex: flagIndex)
-                        .foregroundColor(isActive ? .yellow : .white)
-                        .offset(x: stemInfo.mainStemX + 2, y: stemInfo.stemTop - CGFloat(flagIndex * 8))
-                }
-            }
-            
-            // Drum symbols with individual stems (for single notes or connection to main stem)
-            ForEach(beat.drums, id: \.self) { drum in
-                ZStack {
-                    let drumYOffset = drum.notePosition.yOffset
-                    
-                    // Individual stem for single notes or short connection to main stem
-                    if beat.interval.needsStem {
-                        if !stemInfo.hasConnectedStem {
-                            if isBeamed {
-                                // Beamed stem - calculate height to reach beam position
-                                let beamY: CGFloat = -60 // Base beam position relative to center staff line
-                                let stemHeight = abs(drumYOffset - beamY)
-                                let stemCenterY = (drumYOffset + beamY) / 2
-                                
-                                Rectangle()
-                                    .frame(width: 2, height: stemHeight)
-                                    .foregroundColor(isActive ? .yellow : .white)
-                                    .offset(x: 7, y: stemCenterY)
-                            } else {
-                                // Individual stem for single note (not beamed)
-                                Rectangle()
-                                    .frame(width: 2, height: 75)
-                                    .foregroundColor(isActive ? .yellow : .white)
-                                    .offset(x: 7, y: drumYOffset - 37.5)
-                            }
-                        } else {
-                            // Short connector to main stem (only if note is not at the extreme ends)
-                            let isTopNote = drumYOffset == beat.drums.map { $0.notePosition.yOffset }.min()
-                            let isBottomNote = drumYOffset == beat.drums.map { $0.notePosition.yOffset }.max()
-                            
-                            if !isTopNote && !isBottomNote {
-                                // Short horizontal connector to main stem
-                                Rectangle()
-                                    .frame(width: 5, height: 2)
-                                    .foregroundColor(isActive ? .yellow : .white)
-                                    .offset(x: 4.5, y: drumYOffset)
-                            }
-                        }
-                    }
-                    
-                    // Flags for individual notes (only if not using connected stem and not beamed)
-                    if !stemInfo.hasConnectedStem && beat.interval.needsFlag && !isBeamed {
-                        ForEach(0..<beat.interval.flagCount, id: \.self) { flagIndex in
-                            FlagView(flagIndex: flagIndex)
-                                .foregroundColor(isActive ? .yellow : .white)
-                                .offset(x: 9, y: drumYOffset - 67.5 - CGFloat(flagIndex * 8))
-                        }
-                    }
-                    
-                    // Drum symbol
-                    Text(drum.symbol)
-                        .font(.system(size: 20, weight: .bold))
-                        .foregroundColor(isActive ? .yellow : .white)
-                        .offset(x: 0, y: drumYOffset)
-                }
-            }
-        }
-        .frame(width: 30, height: GameplayLayout.staffHeight)
-    }
-}
-
-// MARK: - Stem Connection Info
-private struct StemConnectionInfo {
-    let hasConnectedStem: Bool
-    let stemTop: CGFloat
-    let stemBottom: CGFloat
-    let mainStemX: CGFloat
-}
-
-// MARK: - Flag View
-struct FlagView: View {
-    let flagIndex: Int
-    
-    var body: some View {
-        Path { path in
-            path.move(to: CGPoint(x: 0, y: 0))
-            path.addCurve(to: CGPoint(x: 8, y: 4), 
-                         control1: CGPoint(x: 4, y: -2), 
-                         control2: CGPoint(x: 6, y: 2))
-            path.addCurve(to: CGPoint(x: 0, y: 8), 
-                         control1: CGPoint(x: 6, y: 6), 
-                         control2: CGPoint(x: 4, y: 10))
-            path.closeSubpath()
-        }
-        .fill(Color.white)
-        .frame(width: 8, height: 8)
-    }
-}
-
-// MARK: - Drum Clef Symbol
-struct DrumClefSymbol: View {
-    var body: some View {
-        VStack(spacing: 4) {
-            // Top rectangle
-            Rectangle()
-                .frame(width: 12, height: 8)
-            
-            // Middle rectangle
-            Rectangle()
-                .frame(width: 12, height: 8)
-            
-            // Bottom rectangle
-            Rectangle()
-                .frame(width: 12, height: 8)
-        }
-        .frame(width: 12, height: 32)
-    }
-}
-
-// MARK: - Time Signature Symbol
-struct TimeSignatureSymbol: View {
-    let timeSignature: TimeSignature
-    
-    var body: some View {
-        VStack(spacing: 2) {
-            // Top number (beats per measure)
-            Text("\(timeSignature.beatsPerMeasure)")
-                .font(.system(size: 18, weight: .bold, design: .serif))
-                .foregroundColor(.white)
-            
-            // Bottom number (note value)
-            Text("\(timeSignature.noteValue)")
-                .font(.system(size: 18, weight: .bold, design: .serif))
-                .foregroundColor(.white)
-        }
-        .frame(width: 25, height: 50)
     }
 }
